Rename package enriched to enrichment
